refactor file location; add more testfiles
