--- .geheim/secrets.geheim.yaml
@@ -1,6 +1,0 @@
----
-akey: a secret
-anArray:
-  - with secrets
-aMap:
-  akey: alsoWithSecrets
