+++ Makefile
@@ -1,9 +1,15 @@
 usage:
 	@echo "Usage:"
 	@echo "	make usage (default)"
+	@echo "	make build_and_test"
 	@echo "	make test_all"
 	@echo "	make test_cov"
 	@echo "	make release_binary_github"
+
+build_and_test:
+	@go build .
+	@./geheim
+	@rm -rf geheim
 
 test_all:
 	@go test -coverprofile=coverage.out ./...
