+++ packages/encrypt/main.go
@@ -5,7 +5,7 @@ import (
 	"encoding/hex"
 	"fmt"
 	"os"
-	"treuzedev/geheim/shared"
+	"treuzedev/geheim/packages/shared"
 )
 
 // TODO: go routines ??
