+++ packages/shared/check-error.go
@@ -5,6 +5,11 @@ package shared
 // TODO: exit gracefully?
 func CheckError(e error) {
 	if e != nil {
-		panic(e)
+		switch e.Error() {
+		case "encoding/hex: invalid byte: U+0000":
+			return
+		default:
+			panic(e)
+		}
 	}
 }
