+++ testfiles/config.json
@@ -1,0 +1,9 @@
+{
+    "auths": {
+      "quay.io": {
+        "email": "",
+        "auth": "ZGNjK3NvbGRldl9wYXJ0bmVyXzEyODQ3NGp2YnNqbjpudmlkd2RiY29oc2JkMjU4NDVka3NiZGIK"
+      }
+    }
+  }
+
