+++ testfiles/known_hosts
@@ -1,0 +1,2 @@
+|1|fOlEoLFkbR2mwBcXi2Mze/u5coA=|lggMyqI62F4GT2Vo6IQkDiTmgo4= ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFSMqzJeV9rUzU4kWitGjeR4PWSa29SPqJ1fVkhtj3Hw9xjLVXVYrU9QlYWrOLXBpQ6KWjbjTDTdDkoohFzgbEY=
+|1|6uxbq+2RqAtaKZu5YP7H/pW/24A=|uJ0higXHYrVcRSNCWfnp7Jd5EIM= ecdsa-sha2-nistp256 AAAAE2VjZHNhLXNoYTItbmlzdHAyNTYAAAAIbmlzdHAyNTYAAABBBFSMqzJeV9rUzU4kWitGjeR4PWSa29SPqJ1fVkhtj3Hw9xjLVXVYrU9QlYWrOLXBpQ6KWjbjTDTdDkoohFzgbEY=
